feat: add to legacy
