+++ test/usage.test.ts
@@ -1,6 +1,5 @@
 import * as cdk from 'aws-cdk-lib';
-import { App } from 'aws-cdk-lib';
-import { StackProps } from 'aws-cdk-lib/core/lib/stack';
+import { App, StackProps } from 'aws-cdk-lib';
 import { Construct } from 'constructs';
 import { CdkExpressPipeline, ExpressStack, ExpressStage, ExpressWave } from '../src';
 import { CdkExpressPipelineLegacy } from '../src/cdk-express-pipeline-legacy';
@@ -138,7 +137,110 @@ describe('CdkExpressPipelineLegacy', () => {
     ]);
 
     expect(wave1Stage1StackA.stackName).toBe('MyWave1Stage1StackA');
+  });
 
+  describe('Correct Stack Highlight for deployment', () => {
+    const app = new App();
+    /* === Wave 1 === */
+    const wave1Stage1StackA = new cdk.Stack(app, 'Wave1Stage1StackA');
+    const wave1Stage1StackB = new cdk.Stack(app, 'Wave1Stage1StackB');
+    const wave1Stage1StackC = new cdk.Stack(app, 'Wave1Stage1StackC');
+    wave1Stage1StackB.addDependency(wave1Stage1StackA);
+
+    /* --- Wave 1, Stage 2 --- */
+    const wave1Stage2StackD = new cdk.Stack(app, 'Wave1Stage2StackD');
+
+    /* === Wave 2 === */
+    const wave2Stage1StackE = new cdk.Stack(app, 'Wave2Stage1StackE');
+    const wave2Stage1StackF = new cdk.Stack(app, 'Wave2Stage1StackF');
+    wave2Stage1StackF.addDependency(wave2Stage1StackE);
+
+    const waves = [
+      {
+        id: 'Wave1',
+        stages: [
+          {
+            id: 'Stage1',
+            stacks: [wave1Stage1StackA, wave1Stage1StackB, wave1Stage1StackC],
+          },
+          {
+            id: 'Stage2',
+            stacks: [wave1Stage2StackD],
+          },
+        ],
+      },
+      {
+        id: 'Wave2',
+        stages: [
+          {
+            id: 'Stage1',
+            stacks: [wave2Stage1StackE, wave2Stage1StackF],
+          },
+        ],
+      },
+    ];
+
+    const pipeline = new CdkExpressPipelineLegacy();
+    const testArray = [
+      {
+        pattern: '**',
+        expected: '\n' +
+          'ORDER OF DEPLOYMENT\n' +
+          '🌊 Waves  - Deployed sequentially, one after another.\n' +
+          '🏗️ Stages - Deployed in parallel, all stages within a wave are deployed at the same time.\n' +
+          '📦 Stacks - Deployed after their dependant stacks within the stage.\n' +
+          '           - Lines prefixed with a pipe (|) indicate stacks matching the CDK pattern.\n' +
+          '\n' +
+          '  🌊 Wave1\n' +
+          '    🏗️ Stage1\n' +
+          '|     📦 Wave1Stage1StackA\n' +
+          '|     📦 Wave1Stage1StackB\n' +
+          '|     📦 Wave1Stage1StackC\n' +
+          '    🏗️ Stage2\n' +
+          '|     📦 Wave1Stage2StackD\n' +
+          '  🌊 Wave2\n' +
+          '    🏗️ Stage1\n' +
+          '|     📦 Wave2Stage1StackE\n' +
+          '|     📦 Wave2Stage1StackF\n',
+      },
+      {
+        pattern: 'Wave1Stage1*',
+        expected: '\n' +
+          'ORDER OF DEPLOYMENT\n' +
+          '🌊 Waves  - Deployed sequentially, one after another.\n' +
+          '🏗️ Stages - Deployed in parallel, all stages within a wave are deployed at the same time.\n' +
+          '📦 Stacks - Deployed after their dependant stacks within the stage.\n' +
+          '           - Lines prefixed with a pipe (|) indicate stacks matching the CDK pattern.\n' +
+          '\n' +
+          '  🌊 Wave1\n' +
+          '    🏗️ Stage1\n' +
+          '|     📦 Wave1Stage1StackA\n' +
+          '|     📦 Wave1Stage1StackB\n' +
+          '|     📦 Wave1Stage1StackC\n' +
+          '    🏗️ Stage2\n' +
+          '      📦 Wave1Stage2StackD\n' +
+          '  🌊 Wave2\n' +
+          '    🏗️ Stage1\n' +
+          '      📦 Wave2Stage1StackE\n' +
+          '      📦 Wave2Stage1StackF\n',
+      },
+    ];
+
+    test.each(testArray)('Testing Legacy Pipeline: %s', ({
+      pattern,
+      expected,
+    }) => {
+      let consoleLogSpy = jest.spyOn(console, 'log').mockImplementation();
+      process.env.CDK_CONTEXT_JSON = JSON.stringify({ 'aws:cdk:bundling-stacks': [pattern] });
+
+      pipeline.printWaves(waves);
+
+      process.env.CDK_CONTEXT_JSON = '';
+      const actual = consoleLogSpy.mock.calls.join('\n');
+      consoleLogSpy.mockRestore();
+
+      expect(actual).toBe(expected);
+    });
   });
 });
 
@@ -277,7 +379,167 @@ describe('CdkExpressPipeline', () => {
     expect(stackA.id).toBe('MyWave1_MyStage1_MyStackA');
   });
 
+  describe('Correct Stack Highlight for deployment', () => {
+    const app = new App();
+    /* === Wave 1 === */
+    const wave1 = new ExpressWave('Wave1');
+    /* --- Wave 1, Stage 1--- */
+    const wave1Stage1 = new ExpressStage('Stage1', wave1);
+    const wave1Stage1StackA = new ExpressStack(app, 'StackA', wave1Stage1);
+    const wave1Stage1StackB = new ExpressStack(app, 'StackB', wave1Stage1);
+    new ExpressStack(app, 'StackC', wave1Stage1);
+    wave1Stage1StackB.addExpressDependency(wave1Stage1StackA);
 
-  //TODO: In Demo project tests - Select all Stacks in Wave 1
-  //TODO: In Demo project tests - Select all Stacks in Wave 1 Stage 1
+    /* --- Wave 1, Stage 2 --- */
+    const wave1Stage2 = new ExpressStage('Stage2', wave1);
+    new ExpressStack(app, 'StackD', wave1Stage2);
+
+    /* === Wave 2 === */
+    const wave2 = new ExpressWave('Wave2');
+    /* === Wave 2, Stage 1 === */
+    const wave2Stage1 = new ExpressStage('Stage1', wave2);
+    const wave2Stage1StackH = new ExpressStack(app, 'StackE', wave2Stage1);
+    const wave2Stage1StackI = new ExpressStack(app, 'StackF', wave2Stage1);
+    wave2Stage1StackI.addExpressDependency(wave2Stage1StackH);
+
+    const waves = [wave1, wave2];
+    const pipeline = new CdkExpressPipeline({
+      waves,
+    });
+    const testArray = [
+      {
+        pattern: '**',
+        expected: '\n' +
+          'ORDER OF DEPLOYMENT\n' +
+          '🌊 Waves  - Deployed sequentially, one after another.\n' +
+          '🏗️ Stages - Deployed in parallel, all stages within a wave are deployed at the same time.\n' +
+          '📦 Stacks - Deployed after their dependant stacks within the stage (dependencies shown below them with ↳).\n' +
+          '           - Lines prefixed with a pipe (|) indicate stacks matching the CDK pattern.\n' +
+          '\n' +
+          '| 🌊 Wave1\n' +
+          '|   🏗️ Stage1\n' +
+          '|     📦 StackA (Wave1_Stage1_StackA)\n' +
+          '|     📦 StackB (Wave1_Stage1_StackB)\n' +
+          '|        ↳ StackA\n' +
+          '|     📦 StackC (Wave1_Stage1_StackC)\n' +
+          '|   🏗️ Stage2\n' +
+          '|     📦 StackD (Wave1_Stage2_StackD)\n' +
+          '| 🌊 Wave2\n' +
+          '|   🏗️ Stage1\n' +
+          '|     📦 StackE (Wave2_Stage1_StackE)\n' +
+          '|     📦 StackF (Wave2_Stage1_StackF)\n' +
+          '|        ↳ StackE\n',
+      },
+      {
+        pattern: 'Wave1_Stage1_*',
+        expected: '\n' +
+          'ORDER OF DEPLOYMENT\n' +
+          '🌊 Waves  - Deployed sequentially, one after another.\n' +
+          '🏗️ Stages - Deployed in parallel, all stages within a wave are deployed at the same time.\n' +
+          '📦 Stacks - Deployed after their dependant stacks within the stage (dependencies shown below them with ↳).\n' +
+          '           - Lines prefixed with a pipe (|) indicate stacks matching the CDK pattern.\n' +
+          '\n' +
+          '| 🌊 Wave1\n' +
+          '|   🏗️ Stage1\n' +
+          '|     📦 StackA (Wave1_Stage1_StackA)\n' +
+          '|     📦 StackB (Wave1_Stage1_StackB)\n' +
+          '|        ↳ StackA\n' +
+          '|     📦 StackC (Wave1_Stage1_StackC)\n' +
+          '    🏗️ Stage2\n' +
+          '      📦 StackD (Wave1_Stage2_StackD)\n' +
+          '  🌊 Wave2\n' +
+          '    🏗️ Stage1\n' +
+          '      📦 StackE (Wave2_Stage1_StackE)\n' +
+          '      📦 StackF (Wave2_Stage1_StackF)\n' +
+          '         ↳ StackE\n',
+      },
+      {
+        pattern: 'Wave1_Stage1*',
+        expected: '\n' +
+          'ORDER OF DEPLOYMENT\n' +
+          '🌊 Waves  - Deployed sequentially, one after another.\n' +
+          '🏗️ Stages - Deployed in parallel, all stages within a wave are deployed at the same time.\n' +
+          '📦 Stacks - Deployed after their dependant stacks within the stage (dependencies shown below them with ↳).\n' +
+          '           - Lines prefixed with a pipe (|) indicate stacks matching the CDK pattern.\n' +
+          '\n' +
+          '| 🌊 Wave1\n' +
+          '|   🏗️ Stage1\n' +
+          '|     📦 StackA (Wave1_Stage1_StackA)\n' +
+          '|     📦 StackB (Wave1_Stage1_StackB)\n' +
+          '|        ↳ StackA\n' +
+          '|     📦 StackC (Wave1_Stage1_StackC)\n' +
+          '    🏗️ Stage2\n' +
+          '      📦 StackD (Wave1_Stage2_StackD)\n' +
+          '  🌊 Wave2\n' +
+          '    🏗️ Stage1\n' +
+          '      📦 StackE (Wave2_Stage1_StackE)\n' +
+          '      📦 StackF (Wave2_Stage1_StackF)\n' +
+          '         ↳ StackE\n',
+      },
+      {
+        pattern: 'Wave2_*',
+        expected: '\n' +
+          'ORDER OF DEPLOYMENT\n' +
+          '🌊 Waves  - Deployed sequentially, one after another.\n' +
+          '🏗️ Stages - Deployed in parallel, all stages within a wave are deployed at the same time.\n' +
+          '📦 Stacks - Deployed after their dependant stacks within the stage (dependencies shown below them with ↳).\n' +
+          '           - Lines prefixed with a pipe (|) indicate stacks matching the CDK pattern.\n' +
+          '\n' +
+          '  🌊 Wave1\n' +
+          '    🏗️ Stage1\n' +
+          '      📦 StackA (Wave1_Stage1_StackA)\n' +
+          '      📦 StackB (Wave1_Stage1_StackB)\n' +
+          '         ↳ StackA\n' +
+          '      📦 StackC (Wave1_Stage1_StackC)\n' +
+          '    🏗️ Stage2\n' +
+          '      📦 StackD (Wave1_Stage2_StackD)\n' +
+          '| 🌊 Wave2\n' +
+          '|   🏗️ Stage1\n' +
+          '|     📦 StackE (Wave2_Stage1_StackE)\n' +
+          '|     📦 StackF (Wave2_Stage1_StackF)\n' +
+          '|        ↳ StackE\n',
+      },
+      {
+        pattern: 'Wave1_Stage1_StackB',
+        expected: '\n' +
+          'ORDER OF DEPLOYMENT\n' +
+          '🌊 Waves  - Deployed sequentially, one after another.\n' +
+          '🏗️ Stages - Deployed in parallel, all stages within a wave are deployed at the same time.\n' +
+          '📦 Stacks - Deployed after their dependant stacks within the stage (dependencies shown below them with ↳).\n' +
+          '           - Lines prefixed with a pipe (|) indicate stacks matching the CDK pattern.\n' +
+          '\n' +
+          '| 🌊 Wave1\n' +
+          '|   🏗️ Stage1\n' +
+          '      📦 StackA (Wave1_Stage1_StackA)\n' +
+          '|     📦 StackB (Wave1_Stage1_StackB)\n' +
+          '|        ↳ StackA\n' +
+          '      📦 StackC (Wave1_Stage1_StackC)\n' +
+          '    🏗️ Stage2\n' +
+          '      📦 StackD (Wave1_Stage2_StackD)\n' +
+          '  🌊 Wave2\n' +
+          '    🏗️ Stage1\n' +
+          '      📦 StackE (Wave2_Stage1_StackE)\n' +
+          '      📦 StackF (Wave2_Stage1_StackF)\n' +
+          '         ↳ StackE\n',
+      },
+    ];
+
+    test.each(testArray)('Testing: %s', ({
+      pattern,
+      expected,
+    }) => {
+      let consoleLogSpy = jest.spyOn(console, 'log').mockImplementation();
+      process.env.CDK_CONTEXT_JSON = JSON.stringify({ 'aws:cdk:bundling-stacks': [pattern] });
+
+      pipeline.printWaves(waves);
+
+      process.env.CDK_CONTEXT_JSON = '';
+      const actual = consoleLogSpy.mock.calls.join('\n');
+      consoleLogSpy.mockRestore();
+
+      expect(actual).toBe(expected);
+    });
+  });
+
+
 });
